start: Starting the development of the 'UseDev' project, an e-commerce focused on the geek theme, setting up the application structure by importing the static files and images from the project. Created the main file 'index.html' with the site structure, added the global styles file 'styles.css', and created specific style files for each section of the site within the 'styles' folder. Initially applying styles to the application header
Starting the construction of the 'UseDev' homepage by developing the application through the creation of the 'index.html' file containing the initial HTML structure of the application. Imported images and icons into the 'assets' folder, separating icon and product images from the other images. Created the 'styles.css' file in the project root containing global styling variables and rules, and added the import of 'styles.css' into the 'index.html'. Created the 'styles' folder and the specific style files for each section of the site, also performing the import in the 'index.html' file. Created the 'README.md' file with a project description. Finally, added the header structure within the <body> tag in 'index.html' and applied header styles through the 'header.css' file